+++ examples/strapi/config/database.ts
@@ -4,7 +4,7 @@ export default ({ env }) => ({
   connection: {
     client: 'sqlite',
     connection: {
-      filename: path.join(__dirname, '..', 'data', 'data.db'),
+      filename: env('DATABASE_FILENAME', path.resolve(process.cwd(), 'data', 'data.db')),
     },
     useNullAsDefault: true,
   },