+++ examples/strapi/config/plugins.ts
@@ -1,5 +1,5 @@
 export default {
-  'image-optimize': {
+  'next-image': {
     enabled: true,
     config: {
       deviceSizes: [640, 750, 828, 1080, 1200, 1920, 2048, 3840],