+++ packages/strapi-plugin-next-image/server/src/routes/content-api.ts
@@ -4,7 +4,7 @@ export default {
     {
       method: 'GET',
       path: '/',
-      handler: 'image-optimize.optimize',
+      handler: 'next-image.optimize',
       config: {
         auth: false,
         policies: [],